pyproject.toml
[project]
authors = [{ name = "Etienne H", email = "et@hossack.me" }]
license = { file = "LICENSE" }
requires-python = ">=3.9,<4.0"
dependencies = [
    "XlsxWriter<4.0.0,>=3.0.1",
    "requests<3.0.0,>=2.28.2",
    "typer>=0.16.0",
    "types-requests>=2.32.4.20250611",
    "python-dotenv>=1.0.0",
]
name = "ridewithgps-to-cuesheet"
version = "0.4.0"
description = "Convert RideWithGPS maps to BC Randonneurs cuesheets"
readme = "README.md"

[project.scripts]
ridewithgps-to-cuesheet = "ridewithgps_to_cuesheet.cli:cli"
rwgps-cuesheet = "ridewithgps_to_cuesheet.cli:cli"

[dependency-groups]
dev = ["mypy>=1.16.1", "poethepoet>=0.34.0", "ruff<1.0.0,>=0.0.263"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ridewithgps_to_cuesheet"]

[tool.ruff]
exclude = [".git", "__pycache__", ".mypy_stubs", ".venv"]
line-length = 120

[tool.ruff.lint]
select = ["C9", "E", "F", "I", "W"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D106", "D203"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
mypy_path = ".mypy_stubs"
exclude = [".mypy_stubs"]
show_error_codes = true
explicit_package_bases = true
show_column_numbers = true

[tool.poe.tasks]
convert-route = "ridewithgps-to-cuesheet --verbose"
lint = "ruff check ."
format-check = "ruff format --check ."
format = "ruff format ."
types-check = "mypy src/ridewithgps_to_cuesheet"
validate = ["lint", "types-check", "format-check"]
